+++ lib/hq/graphql/ext/object_extensions.rb
@@ -71,7 +71,7 @@ module HQ
 
             case association.macro
             when :has_many
-              field name, [type], null: false, klass: model_name do
+              field name, [type], null: ::HQ::GraphQL.nullable_associations?, klass: model_name do
                 if ::HQ::GraphQL.use_experimental_associations?
                   extension FieldExtension::PaginatedArguments, klass: association_klass
                   extension FieldExtension::PaginatedLoader, klass: klass, association: name, internal_association: internal_association
@@ -81,11 +81,11 @@ module HQ
                 instance_eval(&block) if block
               end
             when :has_one
-              field name, type, null: !auto_nil || !has_presence_validation?(association), klass: model_name do
+              field name, type, null: ::HQ::GraphQL.nullable_associations? || !auto_nil || !has_presence_validation?(association), klass: model_name do
                 extension FieldExtension::AssociationLoaderExtension, klass: klass
               end
             else
-              field name, type, null: !auto_nil || !association_required?(association), klass: model_name do
+              field name, type, null: ::HQ::GraphQL.nullable_associations? || !auto_nil || !association_required?(association), klass: model_name do
                 extension FieldExtension::AssociationLoaderExtension, klass: klass
               end
             end